// src/app/shared/services/api/model/get-project-by-id-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetProjectByIdResponse { 
    id?: number;
    name?: string;
    startDate?: string;
    endDate?: string;
}